fix: eliminate all 6 pre-existing compiler warnings + logic fixes (code audit)
Warnings eliminated (0 warnings total after this commit):
- C4267 x3 (detection_controller.cpp): std::set<int> usedObjectIds
  → changed to std::set<size_t> to match objIdx loop variable type
- C4267 x2 (debug_panel.cpp): image.step / rgb.step in QImage constructor
  → added static_cast<int>() for cv::Mat::step (size_t) conversion
- C4701 x1 (video_display.cpp): uninitialized aspectMode variable
  → added default initializer = Qt::KeepAspectRatio

Logic fixes:
- counting_method_panel: add resetTrendChart() method (clears trend chart history)
- main_window::onResetCount(): call resetTrendChart() so trend chart is cleared
  when user resets packaging count (previously history accumulated indefinitely)

UI_UX_TODO.md:
- Add "代碼品質" section with audit findings (fixed + backlog)
- Document roiWidth design gap and vibrator TODO
- Add 5 potential optimization backlog items

+++ basler_cpp/src/core/detection_controller.cpp
@@ -928,8 +928,8 @@ namespace basler
     void DetectionController::updateObjectTracks(const std::vector<DetectedObject> &objects)
     {
         std::map<int, ObjectTrack> newTracks;
-        std::set<int> usedTrackIds;
-        std::set<int> usedObjectIds;
+        std::set<int>    usedTrackIds;
+        std::set<size_t> usedObjectIds;  // 與 objIdx(size_t) 型別一致，消除 C4267 警告
 
         // 第一階段：為現有追蹤更新速度和預測
         for (auto &[trackId, track] : m_objectTracks)
